moving generated clients to be part of the test suite
